refactor: 파일 구조 변경 상수를 사용부로 이관 및 lotto를 domain하위로 이동
