--- src/js/constants/store.js
@@ -1,9 +1,0 @@
-export const COMMIT = {
-  SET_AMOUNT: 'SET_AMOUNT',
-  SET_LOTTO_LIST: 'SET_LOTTO_LIST',
-  SET_IS_VISIBLE_LOTTOS: 'SET_IS_VISIBLE_LOTTOS',
-};
-
-export const DISPATCH = {
-  MAKE_LOTTO_LIST: 'MAKE_LOTTO_LIST',
-};
