+++ src/js/domain/lotto/constants.js
