+++ src/js/domain/lotto/validateLotto.js
@@ -1,4 +1,4 @@
-import { LOTTO_LENGTH, LOTTO_MAX_NUMBER, LOTTO_MIN_NUMBER } from '../constants/lotto.js';
+import { LOTTO_LENGTH, LOTTO_MAX_NUMBER, LOTTO_MIN_NUMBER } from './constants.js';
 
 export function validateLottoLength(lotto) {
   return lotto.length !== LOTTO_LENGTH;
